DB Columns Modified

+++ database/migrations/2026_02_06_064329_create_games_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('home_team_id')->constrained('teams')->onDelete('cascade');
             $table->foreignId('away_team_id')->constrained('teams')->onDelete('cascade');
-            $table->integer('home_team_score')->nullable();
-            $table->integer('away_team_score')->nullable();
-            $table->integer('week')->index(); // 1, 2, 3, 4, 5, 6
+            $table->unsignedTinyInteger('home_team_score')->nullable();
+            $table->unsignedTinyInteger('away_team_score')->nullable();
+            $table->unsignedTinyInteger('week')->index(); // 1, 2, 3, 4, 5, 6
             $table->boolean('is_played')->default(false)->index();
             $table->timestamps();
         });
